Added some features for customizing colors
Added methods to colorize the control. The main color of the control can be set with setColor(); If there is text in the control the color of the text can be set with setTextColor() and if there is a symbol the symbol color can be set with setSymbolColor().
The Regulator control also has setBarColor().